Trying out a new Route system.
The goal here is make the routes type-safe.  This means that if a route
exist in the runtime code, that route will be populated on the server,
or the server will throw an error.  Routes are therefore safe to
reference in the html code.  This is usefull because the most common
problem with static sites are broken links. With this setup, it will
newer happen.

Furthermore, it has some other benifits:

1) The server can do custom url actions. For example add
  a base url or it could hash the static files and add them
  as a suffix.

2) It could be highly parallisable.

3) It has total control of all routes on the servers, this means
   it will be able to generate a sitemap trivially.
   (https://github.com/srid/zulip-archive/issues/10)